feature/dynamic_resourceDoc: extract compile function to DynamicUtil
