+++ obp-api/src/main/scala/code/api/util/DynamicUtil.scala
@@ -1,0 +1,49 @@
+package code.api.util
+import net.liftweb.common.{Box, Failure}
+
+import java.util.concurrent.ConcurrentHashMap
+import scala.reflect.runtime.universe
+import scala.reflect.runtime.universe.runtimeMirror
+import scala.tools.reflect.{ToolBox, ToolBoxError}
+
+object DynamicUtil {
+
+  val toolBox: ToolBox[universe.type] = runtimeMirror(getClass.getClassLoader).mkToolBox()
+
+  // code -> dynamic method function
+  // the same code should always be compiled once, so here cache them
+  private val dynamicCompileResult = new ConcurrentHashMap[String, Box[Any]]()
+  /**
+   * Compile scala code
+   * toolBox have bug that first compile fail, second or later compile success.
+   * @param code
+   * @return compiled Full[function|object|class] or Failure
+   */
+  def compileScalaCode[T](code: String): Box[T] = {
+    val compiledResult: Box[Any] = dynamicCompileResult.computeIfAbsent(code, _ => {
+      val tree = try {
+        toolBox.parse(code)
+      } catch {
+        case e: ToolBoxError =>
+          return Failure(e.message)
+      }
+
+      try {
+        val func: () => Any = toolBox.compile(tree)
+        Box.tryo(func())
+      } catch {
+        case _: ToolBoxError =>
+          // try compile again
+          try {
+            val func: () => Any = toolBox.compile(tree)
+            Box.tryo(func())
+          } catch {
+            case e: ToolBoxError =>
+              Failure(e.message)
+          }
+      }
+    })
+
+    compiledResult.map(_.asInstanceOf[T])
+  }
+}
